Syntax enhancements
