+++ Dyalect/Parser/Model/Model.cs
@@ -135,33 +135,6 @@ namespace Dyalect.Parser.Model
         }
     }
 
-    public sealed class DIf : DNode
-    {
-        public DIf(Location loc) : base(NodeType.If, loc)
-        {
-
-        }
-
-        public DNode Condition { get; set; }
-
-        public DNode True { get; set; }
-
-        public DNode False { get; set; }
-
-        internal override void ToString(StringBuilder sb)
-        {
-            sb.Append("if ");
-            Condition.ToString(sb);
-            True.ToString(sb);
-
-            if (False != null)
-            {
-                sb.Append("else ");
-                False.ToString(sb);
-            }
-        }
-    }
-
     public sealed class DApplication : DNode
     {
         public DApplication(DNode target, Location loc) : base(NodeType.Application, loc)
